cleaned up warnings in mac osx build

+++ com.amd.aparapi.jni/src/cpp/aparapi.cpp
@@ -560,7 +560,7 @@ cl_int KernelArg::setPrimitiveArg(JNIEnv *jenv, int argIdx, int argPos){
          jlong j = jenv->GetStaticLongField(jniContext->kernelClass, fieldID);
          if (config->isVerbose()){
             fprintf(stderr, "clSetKernelArg static primitive long '%s' index=%d pos=%d value=%ld\n",
-                 name, argIdx, argPos, j); 
+                 name, argIdx, argPos, (signed long)j); 
          }
          status = clSetKernelArg(jniContext->kernel, argPos, sizeof(jlong), &j);
       }else{
@@ -568,7 +568,7 @@ cl_int KernelArg::setPrimitiveArg(JNIEnv *jenv, int argIdx, int argPos){
          jlong j = jenv->GetLongField(jniContext->kernelObject, fieldID);
          if (config->isVerbose()){
             fprintf(stderr, "clSetKernelArg primitive long '%s' index=%d pos=%d value=%ld\n",
-                 name, argIdx, argPos, j); 
+                 name, argIdx, argPos, (signed long)j); 
          }
          status = clSetKernelArg(jniContext->kernel, argPos, sizeof(jlong), &j);
       }
@@ -644,10 +644,10 @@ jint writeProfileInfo(JNIContext* jniContext){
          fprintf(jniContext->profileFile, "%d write %s,", pos++, arg->name);
 
          fprintf(jniContext->profileFile, "%lu,%lu,%lu,%lu,",  
-               (arg->arrayBuffer->write.queued - currSampleBaseTime)/1000, 
-               (arg->arrayBuffer->write.submit - currSampleBaseTime)/1000, 
-               (arg->arrayBuffer->write.start - currSampleBaseTime)/1000, 
-               (arg->arrayBuffer->write.end - currSampleBaseTime)/1000);
+               (unsigned long)(arg->arrayBuffer->write.queued - currSampleBaseTime)/1000, 
+               (unsigned long)(arg->arrayBuffer->write.submit - currSampleBaseTime)/1000, 
+               (unsigned long)(arg->arrayBuffer->write.start - currSampleBaseTime)/1000, 
+               (unsigned long)(arg->arrayBuffer->write.end - currSampleBaseTime)/1000);
       }
    }
 
@@ -662,10 +662,10 @@ jint writeProfileInfo(JNIContext* jniContext){
       fprintf(jniContext->profileFile, "%d exec[%d],", pos++, pass);
 
       fprintf(jniContext->profileFile, "%lu,%lu,%lu,%lu,",  
-            (jniContext->exec[pass].queued - currSampleBaseTime)/1000, 
-            (jniContext->exec[pass].submit - currSampleBaseTime)/1000, 
-            (jniContext->exec[pass].start - currSampleBaseTime)/1000, 
-            (jniContext->exec[pass].end - currSampleBaseTime)/1000);
+            (unsigned long)(jniContext->exec[pass].queued - currSampleBaseTime)/1000, 
+            (unsigned long)(jniContext->exec[pass].submit - currSampleBaseTime)/1000, 
+            (unsigned long)(jniContext->exec[pass].start - currSampleBaseTime)/1000, 
+            (unsigned long)(jniContext->exec[pass].end - currSampleBaseTime)/1000);
    }
 
    // 
@@ -684,10 +684,10 @@ jint writeProfileInfo(JNIContext* jniContext){
             fprintf(jniContext->profileFile, "%d read %s,", pos++, arg->name);
 
             fprintf(jniContext->profileFile, "%lu,%lu,%lu,%lu,",  
-                  (arg->arrayBuffer->read.queued - currSampleBaseTime)/1000, 
-                  (arg->arrayBuffer->read.submit - currSampleBaseTime)/1000, 
-                  (arg->arrayBuffer->read.start - currSampleBaseTime)/1000, 
-                  (arg->arrayBuffer->read.end - currSampleBaseTime)/1000);
+                  (unsigned long)(arg->arrayBuffer->read.queued - currSampleBaseTime)/1000, 
+                  (unsigned long)(arg->arrayBuffer->read.submit - currSampleBaseTime)/1000, 
+                  (unsigned long)(arg->arrayBuffer->read.start - currSampleBaseTime)/1000, 
+                  (unsigned long)(arg->arrayBuffer->read.end - currSampleBaseTime)/1000);
          }
       }
    }
@@ -826,7 +826,7 @@ JNI_JAVA(jint, KernelRunner, runKernelJNI)
             return 0L;
          }
          if (config->isVerbose()){
-            fprintf(stderr, "profileBaseTime %lu \n", jniContext->profileBaseTime);
+            fprintf(stderr, "profileBaseTime %lu \n", (unsigned long)jniContext->profileBaseTime);
          }
       }
